New way to handle ephemeral fields while loading.
... actually, the old way should probably be kept... here's an explanation of the old way and the new way before I go and put the old way back in, though:

The problem to be solved is: Sometimes people will want to store and reload records with ephemeral fields (i.e. ones tha don't get stored). The storing side is easy, just exclude them. The reading side is more problematic: How do you put in replacements while reading the data in? We want to be useful/helpful so that if a field *should* be there and isn't, the warning should come at reading time, not as an exception when you try to evaluate the structure. But we also want to allow some fields to sometimes not exist in the data that's read, otherwise we can't really have ephemeral fields.

The old way was to provide a function to produce data field defaults. The problem with this one is, all you get is a function. That is, you can't, for instance, put different values into different instances of the same field of the same record type. On the other hand, it does play well with records with strict fields.

The new way is to provide a function to allow some fields to be excluded. The fields that weren't found in the data and were allowed to be excluded are simply filled with calls to error (with a hopefully useful error message). THe user can then go in and put ephemeral data into those fields (while being careful not to evaluate them)

Really, the two ways are somewhat orthogonal, so it should be OK to support both. So I probably will.